change the type of the Z-axis sign

+++ src/fswAlgorithms/attGuidance/flybyPoint/flybyPoint.h
@@ -38,8 +38,8 @@ public:
     void setTimeBetweenFilterData(double timeBetweenFilterData);
     double getToleranceForCollinearity() const;
     void setToleranceForCollinearity(double toleranceForCollinearity);
-    int64_t getSignOfOrbitNormalFrameVector() const;
-    void setSignOfOrbitNormalFrameVector(int64_t signOfOrbitNormalFrameVector);
+    int getSignOfOrbitNormalFrameVector() const;
+    void setSignOfOrbitNormalFrameVector(int signOfOrbitNormalFrameVector);
 
     ReadFunctor<NavTransMsgPayload>  filterInMsg;               //!< input msg relative position w.r.t. asteroid
     ReadFunctor<EphemerisMsgPayload> asteroidEphemerisInMsg;    //!< input asteroid ephemeris msg
@@ -48,7 +48,7 @@ public:
 private:
     double timeBetweenFilterData = 0;       //!< time between two subsequent reads of the filter information
     double toleranceForCollinearity = 0;            //!< tolerance for singular conditions when position and velocity are collinear
-    int64_t signOfOrbitNormalFrameVector = 1;  //!< Sign of orbit normal vector to complete reference frame
+    int signOfOrbitNormalFrameVector = 1;  //!< Sign of orbit normal vector to complete reference frame
 
     bool firstRead = true;           //!< variable to attest if this is the first read after a Reset
     double f0 = 0;                  //!< ratio between relative velocity and position norms at time of read [Hz]
